+++ src/Network/Network/ChangeLog.md
@@ -19,6 +19,9 @@
 --->
 
 ## Upcoming Release
+* Added EdgeZone parameter support for Azure Firewall
+  * Added `-EdgeZone` parameter to `New-AzFirewall` cmdlet
+  * When EdgeZone is specified, zones property is automatically set to null
 * Onboarded `Microsoft.DataReplication/replicationVaults` to Private Link Common Cmdlets
 
 ## Version 7.25.0